more styles, some fix, mail deliver to file and more

+++ mainapp/models.py
@@ -11,7 +11,7 @@ class Wallet(models.Model):
     name = models.TextField()
     id_user = models.ForeignKey(user_model, on_delete=models.CASCADE)
     categories = models.ManyToManyField('Category')
-    account_balance = models.FloatField()
+    account_balance = models.DecimalField(max_digits=10, decimal_places=2)
 
     def __str__(self):
         return self.name
@@ -27,7 +27,7 @@ class Category(models.Model):
 
 class ExpanseOperation(models.Model):
     title = models.TextField()
-    amount = models.FloatField()
+    amount = models.DecimalField(max_digits=10, decimal_places=2)
     description = models.TextField()
     wallet = models.ForeignKey(Wallet, on_delete=models.CASCADE)
     category = models.ForeignKey(Category, on_delete=models.CASCADE, null=True)
@@ -36,7 +36,7 @@ class ExpanseOperation(models.Model):
 
 class IncomeOperation(models.Model):
     title = models.TextField()
-    amount = models.FloatField()
+    amount = models.DecimalField(max_digits=10, decimal_places=2)
     description = models.TextField()
     wallet = models.ForeignKey(Wallet, on_delete=models.CASCADE, null=True)
     id_user = models.ForeignKey(user_model, on_delete=models.CASCADE)
